requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.13.1
enum-compat==0.0.3
idna==3.3
isodate==0.6.1
lxml==4.7.1
Pillow==8.4.0
platformdirs==2.4.1
pycryptodome==3.12.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
suds-py3==1.4.5.0
uritemplate==4.1.1
urllib3==1.26.8
zeep==4.1.0
